change the type of properties from int  to long

+++ Src/Libraries/2-Application/Application.Share/Dtos/Workspace/Boards/BoardReportDto.cs
@@ -7,7 +7,7 @@ namespace TaskoMask.Application.Share.Dtos.Workspace.Boards
     public class BoardReportDto: TaskReportDto
     {
         [Display(Name = nameof(ApplicationMetadata.CardsCount), ResourceType = typeof(ApplicationMetadata))]
-        public int CardsCount { get; set; }
+        public long CardsCount { get; set; }
 
 
     }
